% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.R
\docType{data}
\name{parks}
\alias{parks}
\title{NPS Park List}
\format{
\subsection{\code{parks}}{

A data frame 595 parks with abbreviated ID and long name:
\describe{
\item{UnitCode}{Park Abbreviated ID}
\item{FullName}{The Park Common Name}
...
}
}
}
\source{
\url{https://irma.nps.gov/NPSpecies/Park}
}
\usage{
parks
}
\description{
A List of All Parks on NPSpecies Site
Report ...
}
\keyword{datasets}
