% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\title{NPS Species List by Park}
\format{
\subsection{\code{species}}{

A data frame 443526 rows with Species Details:
}
}
\source{
\url{https://irma.nps.gov/NPSpecies/Search/SpeciesList}
}
\usage{
species
}
\description{
A List of All Parks on NPSpecies Site
Report ...
}
\keyword{datasets}
